Generify more things
